+++ CHANGELOG.md
@@ -9,6 +9,8 @@ Current
 -------
 ### Fixed:
 
+- [Fix: Presto Limit and the SD build](https://github.com/yahoo/fili/issues/1129)
+
 - [Fix: Bug: dataSourceMetadataService in TestBinderFactory hides instance in AbstractBinderFactory](https://github.com/yahoo/fili/issues/1126)
 
 - [Fix: Another batch of security alerts](https://github.com/yahoo/fili/issues/1123)